+++ GetScaleIndex/GetScaleIndex.cs
@@ -1,6 +1,6 @@
 ﻿using System;
 using System.Collections.Generic;
-using System.Reflection;
+using System.Threading.Tasks;
 
 namespace GetScaleIndex
 {
@@ -300,7 +300,7 @@ namespace GetScaleIndex
             }
 
             //
-            int offset = (index / width * (customScale * (customScale - 1)) * height) + (index * customScale);
+            int offset = (index / width * customScale * (customScale - 1) * height) + (index * customScale);
 
             //
             int[] scaledList = new int[customScale * customScale];
@@ -360,7 +360,7 @@ namespace GetScaleIndex
                 int[] scalingIndexList = GetCustomScalingIndex(index: i, width: width, height: height, customScale: customScale);
 
                 //
-                foreach (var item in scalingIndexList)
+                foreach (int item in scalingIndexList)
                 {
                     //
                     result[item] = list[i];
@@ -370,5 +370,72 @@ namespace GetScaleIndex
             // Returning result
             return result;
         }
+
+        /// <summary>
+        /// Experimental method. [Do not use!]
+        /// </summary>
+        /// <typeparam name="T"></typeparam>
+        /// <param name="array"></param>
+        /// <param name="width"></param>
+        /// <param name="height"></param>
+        /// <param name="customScale"></param>
+        /// <returns></returns>
+        /// <exception cref="NotImplementedException"></exception>
+        public static T[] ApplyScale<T>(T[] array, int width, int height, int customScale)
+        {
+            //
+            T[] result = new T[array.Length * customScale * customScale];
+
+            //
+            for (int i = 0; i < array.Length; i++)
+            {
+                //
+                int[] scalingIndexList = GetCustomScalingIndex(index: i, width: width, height: height, customScale: customScale);
+
+                //
+                for(int j = 0; j < scalingIndexList.Length; j++)
+                {
+                    //
+                    result[scalingIndexList[j]] = array[i];
+                }
+            }
+
+            //
+            return result;
+        }
+
+        /// <summary>
+        /// Experimental method. [Do not use!]
+        /// </summary>
+        /// <typeparam name="T"></typeparam>
+        /// <param name="array"></param>
+        /// <param name="width"></param>
+        /// <param name="height"></param>
+        /// <param name="customScale"></param>
+        /// <param name="parallelOptions"></param>
+        /// <returns></returns>
+        /// <exception cref="NotImplementedException"></exception>
+        public static T[] ApplyScale<T>(T[] array, int width, int height, int customScale, ParallelOptions parallelOptions)
+        {
+            //
+            T[] result = new T[array.Length * customScale * customScale];
+
+            _ = Parallel.For(0, array.Length, parallelOptions: parallelOptions, x =>
+            {
+                //
+                int[] scalingIndexList = GetCustomScalingIndex(index: x, width: width, height: height, customScale: customScale);
+
+                //
+                for (int j = 0; j < scalingIndexList.Length; j++)
+                {
+                    //
+                    result[scalingIndexList[j]] = array[x];
+                }
+            });
+
+            //
+            return result;
+        }
+    }
     }
 }